+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # 0.3.2
+- Unleash Chaos Macro
 - Templates fix.
 - [Removed] Beastform art updater feature and all related settings.
 - [Added] Recovery core feature for Downtime UI: during a short rest, choose one long rest downtime move instead. Can spend 1 Hope to grant the same benefit to an ally. Fully compatible with Efficient (both can be active simultaneously as independent upgrade slots).